feat: support PostUp and PostDown
